pyproject.toml
[tool.poetry]
name = "cobol-converter"
version = "0.1.0"
description = "A simple tool used to convert Cobol into Python and generate unit tests."
authors = ["Gil Fernandes <gil.fernandes@onepointltd.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pyautogen = {git = "https://github.com/microsoft/autogen"}
langchain = "^0.0.335"
pylint = "^3.0.2"
tomli = "^2.0.1"
black = "^23.11.0"


[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
